backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
python-dotenv>=1.0
drf-spectacular>=0.27  # for API schema/docs
django-filter>=24.2    # for query filtering in DRF

# Database drivers
mysqlclient>=2.2
PyMySQL>=1.1
cryptography>=42.0.0